{{cookiecutter.repo_name}}/pyproject.toml
[project]
name = "{{cookiecutter.repo_name}}"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">={{cookiecutter.python_version}}"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
]
authors = [
    {name = "{{cookiecutter.author_name}}"},
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
infra = [
    "aws-cdk-lib==2.199.0",
    "constructs>=10.0.0,<11.0.0",
]




[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]


[project.scripts]
cdk = "scripts.cdk:run_cdk"
start = "scripts.api:start_app"
